+++ app/src/main/java/com/example/roleplaybanking/structures/Game.java
@@ -3,5 +3,5 @@ package com.example.roleplaybanking.structures;
 public class Game {
     public String name;
     public String adminName;
-    public long defaultBalance;
+    public double defaultBalance;
 }
